REFACTOR: move classes
