+++ base/src/main/java/org/aya/concrete/parse/AyaProducer.java
@@ -20,11 +20,8 @@ import org.aya.api.ref.LocalVar;
 import org.aya.api.ref.PreLevelVar;
 import org.aya.concrete.Expr;
 import org.aya.concrete.Pattern;
+import org.aya.concrete.parse.error.*;
 import org.aya.concrete.remark.Remark;
-import org.aya.concrete.resolve.error.BadCounterexampleWarn;
-import org.aya.concrete.resolve.error.PrimDependencyError;
-import org.aya.concrete.resolve.error.RedefinitionError;
-import org.aya.concrete.resolve.error.UnknownPrimError;
 import org.aya.concrete.stmt.*;
 import org.aya.core.def.PrimDef;
 import org.aya.generic.Constants;
@@ -189,12 +186,14 @@ public final class AyaProducer {
     return unreachable(ctx);
   }
 
-  public Tuple2<@NotNull String, OpDecl.@Nullable OpInfo> visitDeclNameOrInfix(@NotNull AyaParser.DeclNameOrInfixContext ctx) {
+  public Tuple2<@NotNull WithPos<String>, OpDecl.@Nullable OpInfo> visitDeclNameOrInfix(@NotNull AyaParser.DeclNameOrInfixContext ctx) {
     var assoc = ctx.assoc();
-    var id = ctx.ID().getText();
-    if (assoc == null) return Tuple.of(id, null);
-    var infix = makeInfix(assoc, id);
-    return Tuple.of(infix.name(), infix);
+    var id = ctx.ID();
+    var txt = id.getText();
+    var pos = sourcePosOf(id);
+    if (assoc == null) return Tuple.of(new WithPos<>(pos, txt), null);
+    var infix = makeInfix(assoc, txt);
+    return Tuple.of(new WithPos<>(pos, infix.name()), infix);
   }
 
   private @NotNull OpDecl.OpInfo makeInfix(@NotNull AyaParser.AssocContext assoc, @NotNull String id) {
@@ -213,12 +212,12 @@ public final class AyaProducer {
     var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
 
     return new Decl.FnDecl(
-      sourcePosOf(ctx.declNameOrInfix()),
+      nameOrInfix._1.sourcePos(),
       sourcePosOf(ctx),
       accessibility,
       modifiers,
       nameOrInfix._2,
-      nameOrInfix._1,
+      nameOrInfix._1.data(),
       visitTelescope(ctx.tele()),
       type(ctx.type(), sourcePosOf(ctx)),
       visitFnBody(ctx.fnBody()),
@@ -557,11 +556,11 @@ public final class AyaProducer {
       body.view().map(ctor -> new WithPos<>(ctor.sourcePos, ctor.ref.name())));
     var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
     var data = new Decl.DataDecl(
-      sourcePosOf(ctx.declNameOrInfix()),
+      nameOrInfix._1.sourcePos(),
       sourcePosOf(ctx),
       accessibility,
       nameOrInfix._2,
-      nameOrInfix._1,
+      nameOrInfix._1.data(),
       visitTelescope(ctx.tele()),
       type(ctx.type(), sourcePosOf(ctx)),
       body,
@@ -569,9 +568,9 @@ public final class AyaProducer {
     );
     return Tuple2.of(data, ctx.OPEN() == null ? ImmutableSeq.empty() : ImmutableSeq.of(
       new Command.Open(
-        sourcePosOf(ctx.declNameOrInfix()),
+        nameOrInfix._1.sourcePos(),
         openAccessibility,
-        new QualifiedID(sourcePosOf(ctx), nameOrInfix._1),
+        new QualifiedID(sourcePosOf(ctx), nameOrInfix._1.data()),
         Command.Open.UseHide.EMPTY
       )
     ));
@@ -594,10 +593,10 @@ public final class AyaProducer {
     var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
     var bind = ctx.bindBlock();
     return new Decl.DataCtor(
-      sourcePosOf(ctx.declNameOrInfix()),
+      nameOrInfix._1.sourcePos(),
       sourcePosOf(ctx),
       nameOrInfix._2,
-      nameOrInfix._1,
+      nameOrInfix._1.data(),
       telescope,
       visitClauses(ctx.clauses()),
       patterns,
@@ -686,13 +685,13 @@ public final class AyaProducer {
     var fields = visitFields(ctx.field());
     checkRedefinition(RedefinitionError.Kind.Field,
       fields.view().map(field -> new WithPos<>(field.sourcePos, field.ref.name())));
-    var nameOrIndex = visitDeclNameOrInfix(ctx.declNameOrInfix());
+    var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
     return new Decl.StructDecl(
-      sourcePosOf(ctx.declNameOrInfix()),
+      nameOrInfix._1.sourcePos(),
       sourcePosOf(ctx),
       accessibility,
-      nameOrIndex._2,
-      nameOrIndex._1,
+      nameOrInfix._2,
+      nameOrInfix._1.data(),
       visitTelescope(ctx.tele()),
       type(ctx.type(), sourcePosOf(ctx)),
       // ctx.ids(),
@@ -714,10 +713,10 @@ public final class AyaProducer {
     var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
     var bind = ctx.bindBlock();
     return new Decl.StructField(
-      sourcePosOf(ctx.declNameOrInfix()),
+      nameOrInfix._1.sourcePos(),
       sourcePosOf(ctx),
       nameOrInfix._2,
-      nameOrInfix._1,
+      nameOrInfix._1.data(),
       telescope,
       type(ctx.type(), sourcePosOf(ctx)),
       Option.of(ctx.expr()).map(this::visitExpr),
@@ -732,10 +731,10 @@ public final class AyaProducer {
     var nameOrInfix = visitDeclNameOrInfix(ctx.declNameOrInfix());
     var bind = ctx.bindBlock();
     return new Decl.StructField(
-      sourcePosOf(ctx.declNameOrInfix()),
+      nameOrInfix._1.sourcePos(),
       sourcePosOf(ctx),
       nameOrInfix._2,
-      nameOrInfix._1,
+      nameOrInfix._1.data(),
       telescope,
       type(ctx.type(), sourcePosOf(ctx)),
       Option.none(),
@@ -824,8 +823,7 @@ public final class AyaProducer {
   public @NotNull Modifier visitFnModifiers(AyaParser.FnModifiersContext ctx) {
     if (ctx.OPAQUE() != null) return Modifier.Opaque;
     if (ctx.INLINE() != null) return Modifier.Inline;
-    if (ctx.PATTERN_KW() != null) return Modifier.Pattern;
-    return unreachable(ctx);
+    /*if (ctx.PATTERN_KW() != null)*/ return Modifier.Pattern;
   }
 
   private @NotNull SourcePos sourcePosOf(ParserRuleContext ctx) {
